pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "memoria"
version = "0.1.0"
description = "Multi-tenant memory service for AI assistants"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}

dependencies = [
    "fastapi>=0.129.0,<0.130.0",
    "uvicorn[standard]>=0.31.0,<0.32.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "pymysql>=1.1.0,<2.0.0",
    "pydantic>=2.5.0,<3.0.0",
    "pydantic-settings>=2.1.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "httpx>=0.27.0,<0.28.0",
    "matrixone-python-sdk>=0.1.16",
    "mcp>=1.26.0,<2.0.0",
    "cryptography>=41.0.0",
    "uuid-utils>=0.10.0",
]

[project.optional-dependencies]
openai-embedding = ["openai>=1.0.0,<2.0.0"]
local-embedding = [
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
]
llm = ["openai>=1.0.0,<2.0.0"]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-xdist>=3.5.0",
    "pytest-cov>=4.1.0",
]

[tool.setuptools.packages.find]
include = ["memoria*"]

[tool.pytest.ini_options]
testpaths = ["memoria/tests", "tests"]
asyncio_mode = "strict"
markers = [
    "slow: marks tests as slow (e2e, subprocess per test)",
    "benchmark: marks performance benchmark tests",
    "local_embedding: marks tests requiring local sentence-transformers",
]
